Just a few tweaks and then...*can it (performantly) host Java ON TOP OF IT!?!*  8)) And .NET and dfkjfldkjfdl allllllll the thingsssss? XD
(And even a lot of what I defined already as "SSP" on top of it too! XDD
)
(Just like how a lot of the C standard library can be written in C X3 )
(But here we're made for running untrusted code, so we need a *very*
clear definition between the core and pure-SSP libraries!  So we may
need to split this and redefine the actual-core XD' )